Moved GasMixture to shared
