+++ Content.Server/Atmos/Components/MapAtmosphereComponent.cs
@@ -1,3 +1,4 @@
+using Content.Shared.Atmos;
 using Content.Shared.Atmos.Components;
 using Content.Shared.Atmos.EntitySystems;
