+++ Content.Server/Atmos/EntitySystems/AtmosphereSystem.API.cs
@@ -1,10 +1,10 @@
 using System.Linq;
 using Content.Server.Atmos.Components;
 using Content.Server.Atmos.Piping.Components;
-using Content.Server.Atmos.Reactions;
 using Content.Server.NodeContainer.NodeGroups;
 using Content.Shared.Atmos;
 using Content.Shared.Atmos.Components;
+using Content.Shared.Atmos.Reactions;
 using Robust.Shared.Map.Components;
 using Robust.Shared.Utility;
