+++ Content.Server/Atmos/EntitySystems/AtmosphereSystem.Map.cs
@@ -1,4 +1,5 @@
 using Content.Server.Atmos.Components;
+using Content.Shared.Atmos;
 using Content.Shared.Atmos.Components;
 using Robust.Shared.GameStates;
 using Robust.Shared.Map.Components;
