+++ Content.Server/Atmos/Piping/Binary/Components/GasPortComponent.cs
@@ -1,3 +1,5 @@
+using Content.Shared.Atmos;
+
 namespace Content.Server.Atmos.Piping.Binary.Components
 {
     [RegisterComponent]
