+++ Content.Server/Atmos/Portable/PortableScrubberSystem.cs
@@ -13,6 +13,7 @@ using Content.Server.NodeContainer.NodeGroups;
 using Content.Server.Audio;
 using Content.Server.Administration.Logs;
 using Content.Server.NodeContainer.EntitySystems;
+using Content.Shared.Atmos;
 using Content.Shared.Database;
 
 namespace Content.Server.Atmos.Portable
