+++ Content.Server/Atmos/Reactions/FrezonProductionReaction.cs
@@ -1,5 +1,6 @@
 ﻿using Content.Server.Atmos.EntitySystems;
 using Content.Shared.Atmos;
+using Content.Shared.Atmos.Reactions;
 using JetBrains.Annotations;
 
 namespace Content.Server.Atmos.Reactions;
