+++ Content.Server/Atmos/Reactions/WaterVaporReaction.cs
@@ -1,5 +1,7 @@
 using Content.Server.Atmos.EntitySystems;
 using Content.Server.Fluids.EntitySystems;
+using Content.Shared.Atmos;
+using Content.Shared.Atmos.Reactions;
 using Content.Shared.Chemistry.Components;
 using Content.Shared.FixedPoint;
 using Content.Shared.Maps;
