+++ Content.Server/Atmos/Serialization/TileAtmosCollectionSerializer.cs
@@ -1,4 +1,5 @@
 ﻿using System.Globalization;
+using Content.Shared.Atmos;
 using Robust.Shared.Serialization;
 using Robust.Shared.Serialization.Manager;
 using Robust.Shared.Serialization.Markdown;
