+++ Content.Server/Destructible/Thresholds/Behaviors/SpawnGasBehavior.cs
@@ -1,4 +1,5 @@
 using Content.Server.Atmos;
+using Content.Shared.Atmos;
 using JetBrains.Annotations;
 
 namespace Content.Server.Destructible.Thresholds.Behaviors;
