+++ Content.Server/Disposal/Unit/Components/DisposalHolderComponent.cs
@@ -1,5 +1,6 @@
 using Content.Server.Atmos;
 using Content.Server.Disposal.Tube.Components;
+using Content.Shared.Atmos;
 using Robust.Shared.Containers;
 
 namespace Content.Server.Disposal.Unit.Components
