+++ Content.Server/Mech/Components/MechAirComponent.cs
@@ -1,4 +1,5 @@
 ﻿using Content.Server.Atmos;
+using Content.Shared.Atmos;
 
 namespace Content.Server.Mech.Components;
