+++ Content.Server/Nutrition/EntitySystems/SmokingSystem.Vape.cs
@@ -13,6 +13,7 @@ using Content.Shared.IdentityManagement;
 using Content.Shared.Interaction;
 using Content.Shared.Nutrition;
 using System.Threading;
+using Content.Shared.Atmos;
 
 /// <summary>
 /// System for vapes
