+++ Content.Server/Power/Generator/GeneratorExhaustGasSystem.cs
@@ -1,5 +1,6 @@
 ﻿using Content.Server.Atmos;
 using Content.Server.Atmos.EntitySystems;
+using Content.Shared.Atmos;
 using Content.Shared.Power.Generator;
 
 namespace Content.Server.Power.Generator;
