+++ Content.Shared/Atmos/GasMixture.cs
@@ -1,13 +1,12 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using Content.Server.Atmos.Reactions;
-using Content.Shared.Atmos;
 using Content.Shared.Atmos.EntitySystems;
+using Content.Shared.Atmos.Reactions;
 using Robust.Shared.Serialization;
 using Robust.Shared.Utility;
 
-namespace Content.Server.Atmos
+namespace Content.Shared.Atmos
 {
     /// <summary>
     ///     A general-purpose, variable volume gas mixture.
