+++ Content.Shared/Atmos/Reactions/GasReactionEnums.cs
@@ -1,0 +1,14 @@
+﻿namespace Content.Shared.Atmos.Reactions;
+
+[Flags]
+public enum ReactionResult : byte
+{
+    NoReaction = 0,
+    Reacting = 1,
+    StopReactions = 2,
+}
+
+public enum GasReaction : byte
+{
+    Fire = 0,
+}
